Apport numerotation bon  de commande et gestion de l'API facture(nouvelle entites,dto,mapper,service)

+++ Gestion-facturation/src/main/java/DTO/BonCommandeDto.java
@@ -5,6 +5,8 @@ import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
+
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -15,9 +17,9 @@ public class BonCommandeDto {
 
     private Long idBonCommande;
     private String numeroBon;
-    private double totalHT;
-    private double Tva; 
-    private double totalTtc;
+    private BigDecimal totalHT;
+    private BigDecimal Tva; 
+    private BigDecimal totalTtc;
     private String statut;
 
    @NotEmpty(message = "Le bon de commande doit contenir au moins une ligne")
